services/data_processing/requirements.txt
onnxruntime-gpu==1.16.1
opencv-python==4.5.5.64
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.27.2
pillow==10.4.0
scikit-image
wheel
annoy
